// apps/web/src/api/__generated__/types/GetPostReportsResponseReportsContentItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */
import type { GetPostReportsResponseReportsContentItemPost } from './GetPostReportsResponseReportsContentItemPost';
import type { GetPostReportsResponseReportsContentItemReporter } from './GetPostReportsResponseReportsContentItemReporter';
import type { GetPostReportsResponseReportsContentItemReviewedBy } from './GetPostReportsResponseReportsContentItemReviewedBy';

export type GetPostReportsResponseReportsContentItem = {
  /** 신고 생성 시각 */
  createdAt: string;
  /** 신고 사유 */
  reason: string;
  post?: GetPostReportsResponseReportsContentItemPost;
  /**
   * 검토 시각
   * @nullable
   */
  reviewedAt?: string | null;
  /**
   * 신고 상세 설명
   * @nullable
   */
  description?: string | null;
  reporter?: GetPostReportsResponseReportsContentItemReporter;
  /** 신고 ID */
  id: number;
  reviewedBy?: GetPostReportsResponseReportsContentItemReviewedBy;
  /**
   * 관리자 처리 메모
   * @nullable
   */
  resolutionNote?: string | null;
  /** 신고 상태 */
  status: string;
};