// apps/web/src/api/__generated__/types/ReviewPostReportRequestResolution.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */

/**
 * 처리 방식
 */
export type ReviewPostReportRequestResolution =
  (typeof ReviewPostReportRequestResolution)[keyof typeof ReviewPostReportRequestResolution];

export const ReviewPostReportRequestResolution = {
  Dismiss: 'DISMISS',
  HidePost: 'HIDE_POST',
} as const;